Bug fix to equals() which was using BigInteger.ZERO rather than BigDecimal.ZERO
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/LEXICON_REFACTOR_BRANCH@3176 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/test/com/bigdata/rdf/internal/XSDDecimalInternalValue.java
@@ -48,7 +48,7 @@ public class XSDDecimalInternalValue<V extends BigdataLiteral> extends
 
     @Override
     public boolean booleanValue() {
-        return value.equals(BigInteger.ZERO) ? false : true;
+        return value.equals(BigDecimal.ZERO) ? false : true;
     }
 
     @Override
